migrate spangroup predictor to its own module
